NNUE
Switches from the existing, texel-tuned evaluation function to one based
on NNUE, a technique used by most leading chess engines.

The net is trained using self-play data generated using tools/datagen,
and using 'bullet'.

More details of the training data used for the net as well as the
network architecture are in NETWORK.md.

tools/tuner is also removed, since it is no longer needed!

bench: 30008362

+++ tools/datagen/src/main.rs
@@ -20,8 +20,8 @@ const DATA_DIR: &str = "datagen";
 
 const DEFAULT_DEPTH: u8 = 8;
 const DEFAULT_STARTING_MOVES: usize = 8;
-const ADJUDICATION_THRESHOLD: i16 = 2000;
-const DRAW_THRESHOLD: i16 = 10;
+const ADJUDICATION_THRESHOLD: i32 = 2000;
+const DRAW_THRESHOLD: i32 = 10;
 const ADJUDICATE_WINS_AFTER: usize = 4;
 const ADJUDICATE_DRAWS_AFTER: usize = 10;
 
@@ -329,7 +329,7 @@ fn random_starting_position(rand: &mut impl Rng) -> Result<Game, ()> {
 }
 
 fn acceptable_starting_position(rand: &mut impl Rng, states: &mut PlayerStates) -> Game {
-    const UNBALANCED_STARTING_EVAL: i16 = 1000;
+    const UNBALANCED_STARTING_EVAL: i32 = 1000;
 
     loop {
         // Skip any games that ended before we got to our starting position
